Fancy centre col scroll
Only wanted the project's column scroll, so used some 'hacks' beautifully involving css 'vh' measurement, to set the centre column to full viewport height and have it scroll. The body as a whole has overscroll: hidden so that it can't scroll. It's not perfect, it doesn't look super, duper great when not fully open, but it work's quite nicely on the most part, so I'll allow it - for now.

On mobile it removes all that scrolling non-sense and just uses normal behaviour.