Fix project cards to use flexbox as actual cards
Trickiest part of this was trying to figure out how the hell SVG's work. Through some testing out, I think if an SVG is a child element of a display:flex container, then it doesn't respect the height or width attribute on the SVG or you have to play around with viewBox. I tried the bandaid fix that always seems to fix everything - wrap it in a <div> - and it worked as expected with sizing. Go figure.